fix: reinstall nuxt graphql request and rename util imparts path
