+++ app/pages/about/policies/index.vue
@@ -6,7 +6,7 @@ import { onMounted } from 'vue'
 
 // HELPERS
 import _get from 'lodash/get'
-import removeTags from '../utils/removeTags'
+import removeTags from '@/utilsremoveTags'
 
 // GQL
 import POLICIES_LIST from '../gql/queries/PoliciesList.gql'
