+++ app/pages/about/student-opportunities/index.vue
@@ -4,8 +4,8 @@ import { NavBreadcrumb, BannerText, SectionWrapper, DividerWayFinder, SectionGen
 
 // HELPERS
 import _get from 'lodash/get'
-import fixUri from '../utils/fixUri'
-import removeTags from '../utils/removeTags'
+import fixUri from '@/utilsfixUri'
+import removeTags from '@/utilsremoveTags'
 
 // GQL
 import STUDENT_OPPORTUNITIES_LIST from '../gql/queries/JobStudentOpportunitiesList.gql'
@@ -138,8 +138,8 @@ const parsedAssociatedTopics = computed(() => {
     <BannerText
       v-if="
         page.buttonUrl &&
-          page.buttonUrl[0] &&
-          page.buttonUrl[0].buttonText
+        page.buttonUrl[0] &&
+        page.buttonUrl[0].buttonText
       "
       class="banner-text"
       :title="page.title"
