+++ app/pages/give/endowments/index.vue
@@ -4,11 +4,11 @@ import { MastheadSecondary, SearchGeneric, BlockCallToAction, SectionGenericList
 
 import _get from 'lodash/get'
 
-import config from '../utils/searchConfig'
-import queryFilterHasValues from '../utils/queryFilterHasValues'
-import parseFilters from '../utils/parseFilters'
-import fixUri from '../utils/fixUri'
-import removeTags from '../utils/removeTags'
+import config from '@/utilssearchConfig'
+import queryFilterHasValues from '@/utilsqueryFilterHasValues'
+import parseFilters from '@/utilsparseFilters'
+import fixUri from '@/utilsfixUri'
+import removeTags from '@/utilsremoveTags'
 
 // GQL
 import ENDOWMENTS_LIST from '../gql/queries/EndowmentList.gql'
@@ -131,12 +131,12 @@ const searchGenericQuery = ref({
 
 watch(() =>
   route.query,
-(newVal, oldVal) => {
-  // console.log('ES newVal, oldVal', newVal, oldVal)
-  searchGenericQuery.value.queryText = route.query.q || ''
-  searchGenericQuery.value.queryFilters = parseFilters(route.query.filters || '')
-  searchES()
-}, { deep: true, immediate: true }
+  (newVal, oldVal) => {
+    // console.log('ES newVal, oldVal', newVal, oldVal)
+    searchGenericQuery.value.queryText = route.query.q || ''
+    searchGenericQuery.value.queryFilters = parseFilters(route.query.filters || '')
+    searchES()
+  }, { deep: true, immediate: true }
 )
 
 async function searchES() {
@@ -278,7 +278,7 @@ onMounted(async () => {
         parsedFeaturedEndowments.length &&
         hits.length == 0 &&
         !noResultsFound
-      "
+        "
       class="section-no-top-margin"
       :section-title="page.featuredEndowments[0].titleGeneral"
       :section-summary="page.featuredEndowments[0].sectionSummary"
@@ -296,7 +296,7 @@ onMounted(async () => {
         parsedFeaturedEndowments.length &&
         hits.length == 0 &&
         !noResultsFound
-      "
+        "
       theme="divider"
     >
       <DividerWayFinder color="about" />
@@ -308,7 +308,7 @@ onMounted(async () => {
         parsedEndowmentsList.length &&
         hits.length == 0 &&
         !noResultsFound
-      "
+        "
       section-title="All Collection Endowments"
     >
       <SectionGenericList :items="parsedEndowmentsList" />
