+++ app/pages/help/[slug].vue
@@ -6,7 +6,7 @@ import { onMounted } from 'vue'
 
 // HELPERS
 import _get from 'lodash/get'
-import removeTags from '../utils/removeTags'
+import removeTags from '@/utilsremoveTags'
 
 // GQL
 import HELP_TOPIC_DETAIL from '../gql/queries/HelpTopicDetail.gql'
