+++ app/pages/help/tutorials/index.vue
@@ -10,10 +10,10 @@ import _get from 'lodash/get'
 import TUTORIALS_LIST from '../gql/queries/TutorialsList.gql'
 
 // HELPERS
-import config from '../utils/searchConfig'
-import getListingFilters from '../utils/getListingFilters'
-import queryFilterHasValues from '../utils/queryFilterHasValues'
-import removeTags from '../utils/removeTags'
+import config from '@/utilssearchConfig'
+import getListingFilters from '@/utilsgetListingFilters'
+import queryFilterHasValues from '@/utilsqueryFilterHasValues'
+import removeTags from '@/utilsremoveTags'
 
 const { $graphql } = useNuxtApp()
