+++ app/pages/visit/events-exhibitions/[slug].vue
@@ -5,7 +5,7 @@ import { NavBreadcrumb, BannerText, BannerHeader, DividerGeneral, SectionCardsWi
 
 // HELPERS
 import _get from 'lodash/get'
-import removeTags from '../utils/removeTags'
+import removeTags from '@/utilsremoveTags'
 
 // GQL
 import EVENT_DETAIL from '../gql/queries/EventDetail.gql'
@@ -644,10 +644,8 @@ const { hasCTA } = useAskALibrarianCTA()
       </SectionWrapper>
 
       <SectionWrapper :section-title="parsedAcknowledgementTitle">
-        <RichText
-          :rich-text-content="page?.exhibition?.acknowledgements[0]?.acknowledgements
-          "
-        />
+        <RichText :rich-text-content="page?.exhibition?.acknowledgements[0]?.acknowledgements
+          " />
       </SectionWrapper>
     </div>
   </main>
