+++ nuxt.config.ts
