+++ package.json
@@ -22,6 +22,7 @@
     "@nuxtjs/sitemap": "^5.3.5",
     "@pinia/nuxt": "^0.5.5",
     "@vue-a11y/skip-to": "^3.0.3",
+    "graphql-tag": "^2.12.6",
     "nuxt": "^4.4.7",
     "vue": "^3.5.12",
     "vue-router": "^4.4.5"
@@ -41,8 +42,8 @@
     "eslint-plugin-cypress": "^2.15.1",
     "lodash": "^4.17.21",
     "netlify-cli": "^17.36.0",
-    "nuxt-graphql-request": "^7.0.5",
     "sass": "^1.66.1",
+    "nuxt-graphql-request": "8.1.1",
     "ucla-library-design-tokens": "^6.0.2"
   },
   "packageManager": "pnpm@11.5.1+sha512.93f7b57422ea7068257235b4c16eb60762eb68e1dc23723199cc739043ea9be2c4143274a399d8c6defa2b1176226d9ca1c4b63482d6200c1a8fbaa78c1d1485",
